requirements.txt
openai
SpeechRecognition
pyttsx3
pyaudio
pyautogui
python-dotenv

